% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbwg-eurocontrol-package.R
\docType{package}
\name{pbwg.eurocontrol-package}
\alias{pbwg.eurocontrol}
\alias{pbwg.eurocontrol-package}
\title{PBWG EUROCONTROL helper package}
\description{
Provides high-level wrappers combining \code{eurocontrol} primitives and
purpose-built SQL for the PBWG ECTL workflows.
}
\author{
\strong{Maintainer}: Quinten GOENS EUROCONTROL PRU \email{quinten.goens@eurocontrol.int}

}
\keyword{internal}
