% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf.R
\name{pbwg_apdf_domestic_prefixes_default}
\alias{pbwg_apdf_domestic_prefixes_default}
\title{ICAO prefixes treated as domestic}
\usage{
pbwg_apdf_domestic_prefixes_default()
}
\value{
Character vector of prefixes (upper-case).
}
\description{
Helper list of ICAO country/region prefixes that define "domestic" APDF
traffic for PBWG reporting.
}
\keyword{internal}
